requirements.txt
browser-use
langchain-google-genai
langchain-openai
python-dotenv
mcp
playwright
splunk-sdk
